+++ packages/compose/src/private-apis.ts
@@ -2,12 +2,12 @@
  * Internal dependencies
  */
 import { lock } from './lock-unlock';
-import subscribeSharedListener from './utils/subscribe-shared-listener';
+import subscribeDelegatedListener from './utils/subscribe-delegated-listener';
 
 /**
  * Private @wordpress/compose APIs.
  */
 export const privateApis = {};
 lock( privateApis, {
-	subscribeSharedListener,
+	subscribeDelegatedListener,
 } );